refactor(workbench): fold the workbench lock into the dev-server registry
The registry, the workbench lock, and process-liveness were three modules for
one concept — the on-disk record of running dev servers plus the single-workbench
lock. The lock already depended on the registry, and both re-derived the same
"OS start time, prune if dead" logic. The lock now lives in the registry module
(one seam: register / list / watch / claim); process-liveness stays its internal
primitive; REGISTRY_VERSION and getRegistryDir stop being exported now that
nothing outside the module needs them. Callers cross one seam instead of three.
