+++ packages/@sanity/cli/src/actions/dev/workbench/__tests__/startWorkbenchDevServer.test.ts
@@ -28,12 +28,10 @@ vi.mock('../writeWorkbenchRuntime.js', () => ({
   writeWorkbenchRuntime: mockWriteWorkbenchRuntime,
 }))
 vi.mock('../../registry/registry.js', () => ({
-  getRegisteredServers: mockGetRegisteredServers,
-  watchRegistry: mockWatchRegistry,
-}))
-vi.mock('../../registry/workbenchLock.js', () => ({
   acquireWorkbenchLock: mockAcquireWorkbenchLock,
+  getRegisteredServers: mockGetRegisteredServers,
   readWorkbenchLock: mockReadWorkbenchLock,
+  watchRegistry: mockWatchRegistry,
 }))
 
 function createMockServer(port = 3333) {
